ml-backend/requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.0

# ML and Data Science
scikit-learn==1.5.2
xgboost==2.1.2
numpy==2.1.3
pandas==2.2.3
joblib==1.4.2

# NLP and Text Processing
nltk==3.9.1
spacy==3.8.2
textstat==0.7.4
python-docx==1.1.2
PyPDF2==3.0.1
pdfplumber==0.11.4

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.27.2

# CORS
fastapi-cors==0.0.6

# Logging and Monitoring
structlog==24.4.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
